// src/app/model/client-profile.model.ts
export class ClientProfile{
    userId: number = 0;
    userCode: string= "";
    firstName: string= "";
    lastName: string= "";
    address: string= "";
    district: string= "";
    districtDisplay: string= "";
    city: string= "";
    cityDisplay: string= "";
    phone: string= "";
    email: string= "";
    birthday: string= "";
    sex: string= "";
    username: string= "";
    password: string= "";
    status: string= "";
    statusDisplay: string= "";
    rightId: number = 0;
    rightDisplay: string= "";
    lastChangePass: string= "";
    type: string= "";
    avatarUrl: string= "";
    userLevel: string= "";
    warehouseReceive: string= "";
    createdDate: string= "";
    changeDate: string= "";
    lockedDate: string= "";
    createdUser: string= "";
    changeUser: string= "";
    failureCount: string= "";
    ward: string = "";

    public ClientProfile(){};
}